refactor(account): rename kebab-case dirs to camelCase (#5291)
