+++ jsapp/js/account/addOns/updateBadge.component.tsx
